Treat license: private as private: true.
This is an alternative means for communicating that packages are
private. Doesn't have publish prevention support from npm, but I've seen
it used around the place and seems a good way to flag proprietary code
i.e.  I often see private: true with a license which can't possibly make
sense for private code.